(1) Fixed an error that compiled fine in Eclipse but not with javac (2) Made id name and number fields private.
--HG--
extra : convert_revision : svn%3A849a83f0-50c2-4045-bdd1-8b1f698b4b18/trunk%4097

+++ jsoar/test/org/jsoar/kernel/io/CycleCountInputTest.java
@@ -74,7 +74,7 @@ public class CycleCountInputTest
         int expected = 1;
         for(Integer i : matches)
         {
-            assertEquals(expected++, i);
+            assertEquals(expected++, i.intValue());
         }
     }
 }
